requirements.txt
# Audio recording & playback
sounddevice>=0.4.6
pygame>=2.5.0
numpy>=1.26.0

# Speech-to-Text
faster-whisper>=1.0.0

# LLM (Doubao via 火山方舟, OpenAI-compatible)
openai>=1.50.0

# TTS (Phase 1 placeholder; Phase 2 will add GPT-SoVITS)
edge-tts>=6.1.12

# Config
python-dotenv>=1.0.0

# Note: On Windows, ctranslate2 wheels bundle CuBLAS/cuDNN — no nvidia-* pip packages needed.
# CUDA GPU inference requires only the NVIDIA display driver (>=520) to be installed.
